Use pytest fixtures across tests, add deprecation warning in reader
